// apps/frontend/src/enums/FontWeight.ts
export enum FontWeight {
  Thin = 'font-weight-thin',
  ExtraLight = 'font-weight-extra-light',
  Light = 'font-weight-light',
  Regular = 'font-weight-regular',
  Medium = 'font-weight-medium',
  SemiBold = 'font-weight-semi-bold',
  Bold = 'font-weight-bold',
  ExtraBold = 'font-weight-extra-bold',
  SuperBold = 'font-weight-super-bold',
}