// apps/frontend/src/enums/Color.ts
export enum Color {
  PRIMARY = 'color-primary',
  SECONDARY = 'color-secondary',
  TERTIARY = 'color-tertiary',
  QUATERNARY = 'color-quaternary',
  QUINARY = 'color-quinary',
  BLACK = 'color-black',
  WHITE = 'color-white',
  ERROR = 'color-error',
  SUCCESS = 'color-success',
  WARNING = 'color-warning',
  INFO = 'color-info',
  GRAY = 'color-gray',
}